Flesh out Evildoer
* explain compilation of `begin'
* explain push, pop, call
* explain ABI for calling C
* explain "extern" in ASM
* discuss asm-interp/io as analog of interp-io
* show examples with asm-interp/io, asm-interp only for "pure" programs
* add peek-byte?
