+++ langs/evildoer/double.c
@@ -1,0 +1,5 @@
+#include <inttypes.h>
+
+int64_t dbl(int64_t x) {
+  return x + x;
+}
