+++ langs/evildoer/life.c
@@ -1,0 +1,5 @@
+#include <inttypes.h>
+
+int64_t meaning(void) {
+  return 42;
+}
