+++ langs/evildoer/test/all.rkt
@@ -11,10 +11,10 @@
 (current-shared? #t)
 
 ;; link with byte.o for IO operations
-(unless (file-exists? "../byte.o")
-  (system "make -C .. byte.o"))
+(unless (file-exists? "../byte-shared.o")
+  (system "make -C .. byte-shared.o"))
 (current-objs
- (list (path->string (normalize-path "../byte.o"))))
+ (list (path->string (normalize-path "../byte-shared.o"))))
 
 (define (test-runner run)
   ;; Abscond examples
@@ -73,7 +73,10 @@
   (check-equal? (run '(eof-object? (read-byte)) "") (cons #t ""))
   (check-equal? (run '(eof-object? (read-byte)) "a") (cons #f ""))
   (check-equal? (run '(begin (write-byte 97) (write-byte 98)) "")
-                (cons (void) "ab")))
+                (cons (void) "ab"))
+
+  (check-equal? (run '(peek-byte) "ab") (cons 97 ""))
+  (check-equal? (run '(begin (peek-byte) (read-byte)) "ab") (cons 97 "")))
 
 (test-runner-io (λ (e s) (interp/io (parse e) s)))
 (test-runner-io (λ (e s)
